+++ nion/instrumentation/test/ScanControl_test.py
@@ -974,6 +974,27 @@ class TestScanControlClass(unittest.TestCase):
             self.assertEqual(stem_controller_.probe_position, probe_graphic.position)
             self.assertEqual(stem_controller_.probe_position, hardware_source._get_last_idle_position_for_test())
 
+    def test_setting_probe_graphic_updates_probe_position_with_multiple_channels_enabled(self):
+        with self.__test_context() as test_context:
+            document_controller = test_context.document_controller
+            document_model = test_context.document_model
+            stem_controller_ = test_context.instrument
+            hardware_source = test_context.hardware_source
+            scan_state_controller = test_context.state_controller
+            stem_controller_.set_probe_position(Geometry.FloatPoint(y=0.5, x=0.5))
+            hardware_source.set_channel_enabled(0, True)
+            hardware_source.set_channel_enabled(1, True)
+            scan_state_controller.handle_positioned_check_box(True)
+            self._acquire_one(document_controller, hardware_source)
+            document_controller.periodic()
+            display_item = document_model.get_display_item_for_data_item(document_model.data_items[0])
+            probe_graphic = display_item.graphics[0]
+            self.assertEqual(stem_controller_.probe_position, probe_graphic.position)
+            probe_graphic.position = Geometry.FloatPoint(y=0.45, x=0.65)
+            document_controller.periodic()
+            self.assertEqual(stem_controller_.probe_position, probe_graphic.position)
+            self.assertEqual(stem_controller_.probe_position, hardware_source._get_last_idle_position_for_test())
+
     def test_acquire_into_empty_scan_controlled_display_panel(self):
         ScanControlPanel.run()
         with self.__test_context() as test_context: